[Feature #162027526] Add auth controller for user to signUp
